Better feedback from database for checking dietary preferences

+++ app/src/main/java/com/info301/mypantryapp/domain/Product.java
@@ -4,7 +4,7 @@ public class Product {
     private String productId;
     private String weight;
     private long barcodeNum;
-    private int shelfLife; // in days, using 10000 if n/a
+    private long shelfLife; // in days, using 10000 if n/a
     private String name;
     private String description;
     private String brand;
@@ -12,7 +12,7 @@ public class Product {
 
 //Domain datatype attribute has been changed, database only takes a string at the moment.
 
-    public Product(String productId, String weight, long barcodeNum, int shelfLife, String name, String description, String brand, String ingredients) {
+    public Product(String productId, String weight, long barcodeNum, long shelfLife, String name, String description, String brand, String ingredients) {
         this.productId = productId;
         this.weight = weight;
         this.barcodeNum = barcodeNum;
@@ -40,7 +40,7 @@ public class Product {
         return barcodeNum;
     }
 
-    public int getShelfLife() {
+    public long getShelfLife() {
         return shelfLife;
     }
